remove factory classes
- collapse HubFactory, IPControllerApp into IPController
- collapse EngineFactory, IPEngineApp into IPEngine
- move apps into `subpackage.app` instead of all together in `.apps`
- add deprecations so old config still works
- eliminate SessionFactory, RegistrationFactory
