+++ ipyparallel/apps/ipclusterapp.py
@@ -1,0 +1,11 @@
+import warnings
+
+warnings.warn(f"{__name__} is deprecated in ipyparallel 7. Use ipyparallel.cluster")
+
+from ipyparallel.cluster.app import IPCluster, main
+
+IPClusterApp = IPCluster
+launch_new_instance = main
+
+if __name__ == "__main__":
+    main()
