+++ ipyparallel/engine/__main__.py
@@ -1,8 +1,4 @@
-def main():
-    from ipyparallel.apps import ipengineapp as app
-
-    app.launch_new_instance()
-
-
 if __name__ == '__main__':
+    from .app import main
+
     main()
